requirements.txt
annotated-types==0.7.0
async-timeout==5.0.0
gpiod==2.2.1
gpiodevice==0.0.5
inky==2.0.0
numpy==2.1.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
redis==5.2.0
ruff==0.7.2
smbus2==0.5.0
spidev==3.6
tenacity==9.0.0
typing_extensions==4.12.2
